requirements.txt
google-genai
huggingface_hub
pillow
python-dotenv
sqlalchemy
gradio
